+++ totalRP3/tools/Locale.lua
@@ -1692,7 +1692,7 @@ We are aware of a current issue on Retail causing **quest item usage from the ob
 
 ## Added
 
-- Added 323 new icons, X images and 138 new music files from patch 9.1.0.
+- Added 323 new icons, 7 images and 138 new music files from patch 9.1.0.
 
 ## Removed